reworked target code to support multiple targets and different target types.
user can now specify a sequence of targets to test (ordered by rule number). The first success satifies the success
criteria for the test and no further tests will be performed on the interface for this period.

Additional test types can be coded and added to the target framework.

configuration has changed as a result of the rework.