fix for bug 5991 and 5715.
The problem was introduced when the start order for WLB and pppoe interfaces changed (must have been sometime pre-kenwood). The
pppoe hook script was checking for the existence of WLB prior to updating the nexthop information for WLB. The fix is to perform the
update without this check and WLB will pick up the changes when it is eventually started.